atk,gio,gtk: Rename file names to reflect type names
